requirements.txt
pytorch-lightning>=2.0.0
torch>=2.0.0
transformers>=4.30.0
datasets>=2.14.0
openai>=1.0.0
kaggle>=1.5.0
vastai>=0.1.0
pyyaml>=6.0
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0
tqdm>=4.65.0
python-dotenv>=1.0.0